Add subindex 1 in PWB_SDO_FAN_CONFIGURATION
Use subindex 1 in PWB_SDO_FAN_CONFIGURATION to return the number of
fans, so that the Charger application can get and use it.

Previously, there was only subindex 0 which represents the configuration
id. If the Charger application wants to infer the fan number, it needs
to define the same structure as the one in PowerBridge. In that case, if
the structure in PowerBridge is changed, the Charger's code needs to be
changed accordingly.

Change-Id: I5504818a0aaf2ce83ddba140db22a3f40efa2df6